requirements.txt
fastapi
uvicorn[standard]
python-docx
openpyxl
python-multipart
